Store port addresses and shape counts as integers
Ports and shape counts were stored as float32 everywhere: to_numpy forced
float32 at creation, and Backend.from_numpy re-cast them to the backend's
floating dtype on every to_backend call, compensated downstream by
astype(int) at gather/scatter sites. This silently corrupts addresses
beyond 2^24, and beyond 256 with a bfloat16 backend.

Ports and shape counts are now int32 end to end:

- HyperEdgeSet.from_dict casts ports to int32 and features to float32;
  integer-valued float ports are still accepted, fractional or NaN ports
  are rejected.
- Backend.from_numpy accepts a PRESERVE_DTYPE sentinel so that to_backend
  never routes integer data (ports, shape counts, feature-name indices)
  through the backend's floating dtype.
- to_numpy preserves dtypes instead of forcing float32.
- GraphShape counts are built as int32; GraphShape.max defaults missing
  classes to 0 instead of -inf to avoid float promotion.
- Compensatory astype(int) casts in gather/scatter call sites are removed.

Benchmarked on a jitted grad step through LocalSumMessagePassingFunction:
no measurable difference in time or peak memory vs float32 ports — the
change is about correctness, not speed.

+++ src/energnn/model/utils.py
@@ -115,7 +115,7 @@ def gather(*, coordinates: jax.Array, addresses: jax.Array) -> jax.Array:
     :param addresses: Integer indices specifying which elements to gather.
     :returns: Gathered elements of the same shape as `addresses`.
     """
-    return coordinates.at[addresses.astype(int)].get(mode="drop", fill_value=0.0)
+    return coordinates.at[addresses].get(mode="drop", fill_value=0.0)
 
 
 def scatter_add(*, accumulator: jax.Array, increment: jax.Array, addresses: jax.Array) -> jax.Array:
@@ -127,4 +127,4 @@ def scatter_add(*, accumulator: jax.Array, increment: jax.Array, addresses: jax.
     :param addresses: Integer indices where increments should be added.
     :returns: Updated accumulator array after adding increments.
     """
-    return accumulator.at[addresses.astype(int)].add(increment, mode="drop")
+    return accumulator.at[addresses].add(increment, mode="drop")
